Make global clamp function due to frequency on being used
#114
